Add toolTipDesc to all item classes and update rules
Added the toolTipDesc property to all item classes to provide tooltip descriptions for each item. Updated GrilledSteak to double max health instead of adding a fixed value. In RuleHandler, added a call to reload the discard pile when the opponent's health reaches zero meaning the game is won to prevent issues with items that affect the whole deck.